api/requirements.txt
beautifulsoup4==4.11.1
fitz==0.0.1.dev2
Flask==2.2.2
gensim==4.1.2
ipython==8.7.0
nlp==0.4.0
nltk==3.7
numpy==1.21.5
pandas==1.2.3
pdf2image==1.16.0
Pillow==9.3.0
pyLDAvis==3.3.1
PyPDF2==2.12.1
pytesseract==0.3.10
requests==2.23.0
spacy==3.4.1
